Clarify docs for toSecond.
I wasn't 100% sure based on the existing docs whether this was
intended to be analagous to toHour, toMinute, etc, or if it just
converted to posix time in seconds. In hindsight, the latter wouldn't
have made much sense, but I was thrown by the facts that:

1. All of the other functions which return a value in a range like this
   say "From 0 to ..", but this one doesn't.
2. The examples don't disambiguate this.

So this patch makes minor improvements to the docs, including the range
for consistency and adding another example that disambiguates.